+++ Addon/ElvUI_mMediaTag/core/options/core.lua
@@ -42,15 +42,6 @@ local function configTable()
 						name = L["Roll Button"],
 						args = {},
 					},
-					chat = {
-						order = 4,
-						type = "group",
-						name = L["Chat Button"],
-						hidden = function()
-							return not E.Retail
-						end,
-						args = {},
-					},
 					keystochat = {
 						order = 5,
 						type = "group",
@@ -309,12 +300,6 @@ local function configTable()
 							},
 						},
 					},
-					-- classcolor = {
-					-- 	order = 3,
-					-- 	type = "group",
-					-- 	name = L["Custom Class colors"],
-					-- 	args = {},
-					-- },
 					roleicons = {
 						order = 4,
 						type = "group",
@@ -352,16 +337,6 @@ local function configTable()
 						order = 8,
 						type = "group",
 						name = L["Datapanels"],
-						childGroups = "tab",
-						args = {},
-					},
-					questicons = {
-						order = 9,
-						type = "group",
-						name = L["Quest Icons"],
-						hidden = function()
-							return not E.Retail
-						end,
 						childGroups = "tab",
 						args = {},
 					},