+++ Addon/ElvUI_mMediaTag/core/options/objectivetracker.lua
@@ -132,8 +132,18 @@ local function configTable()
 						mMT.Modules.ObjectiveTracker:Initialize()
 					end,
 				},
-				header_color = {
+				spacer_1 = {
 					order = 2,
+					type = "description",
+					name = "\n",
+				},
+				info = {
+					order = 3,
+					type = "description", --#FFA7A7
+					name = "|CFFFFA7A7" .. L["Info: The Skin can be affected by other addons if they add a skin for all windows. To fix the problem, the skin must be deactivated in the other addon. This is not a bug of mMT."] .. "|r",
+				},
+				header_color = {
+					order = 4,
 					type = "group",
 					inline = true,
 					name = L["Background Color"],
@@ -183,7 +193,7 @@ local function configTable()
 					},
 				},
 				header_border = {
-					order = 3,
+					order = 5,
 					type = "group",
 					inline = true,
 					name = L["Border"],
@@ -230,7 +240,7 @@ local function configTable()
 					},
 				},
 				header_shadow = {
-					order = 4,
+					order = 6,
 					type = "group",
 					inline = true,
 					name = L["Shadow"],